terminado de compilar (unico problema de casteos => sacamos la flag -Werror), modificamos el main para que tome argumentos del tipo double (Tf y Tr), arreglamos algunos errores de logica, NO HAY LEAKS!, tampoco tiro errores con valgrind. (habia sigsevs pero fueron arreglados)

+++ modelos/laundry.h
@@ -23,7 +23,8 @@ typedef struct _laundry *laundry_t;
  *	Tf:	tiempo medio de fallo de una lavadora operativa
  *	Tr	tiempo medio de reparación de un mecánico
  */
-laundry_t laundry_create (unsigned int Nop, unsigned int Nserv, unsigned int Nmech, unsigned int Tf, unsigned int Tr);
+laundry_t laundry_create (unsigned int Nop, unsigned int Nserv, 
+			   unsigned int Nmech, double Tf, double Tr);
 
 
 /* Destructor del TAD
